+++ kanon/conftest.py
@@ -9,6 +9,8 @@ packagename.test
 
 import os
 
+from hypothesis.strategies._internal.core import tuples
+
 # For Astropy 3.0 and later, we can use the standalone pytest plugin
 try:
     from pytest_astropy_header.display import (PYTEST_HEADER_MODULES,
@@ -47,7 +49,7 @@ def _hypothesis_sexagesimal_strategy():
     from hypothesis.strategies import (builds, decimals, integers, lists,
                                        register_type_strategy, sampled_from)
 
-    from kanon.units import Sexagesimal
+    from kanon.units import Historical, Sexagesimal
 
     settings.register_profile("def", suppress_health_check=(HealthCheck.too_slow,))
     settings.load_profile("def")
@@ -60,6 +62,15 @@ def _hypothesis_sexagesimal_strategy():
         sign=sampled_from((-1, 1))
     )
     register_type_strategy(Sexagesimal, strat)
+
+    strat = builds(
+        Historical,
+        tuples(integers(0, 9), integers(0, 11), integers(0, 29)),
+        lists(integers(0, 59), max_size=2),
+        remainder=decimals(0, 1).filter(lambda x: x != 1),
+        sign=sampled_from((-1, 1))
+    )
+    register_type_strategy(Historical, strat)
 
 # Uncomment the last two lines in this block to treat all DeprecationWarnings as
 # exceptions. For Astropy v2.0 or later, there are 2 additional keywords,